+++ src/utility.jl
@@ -170,15 +170,15 @@ Check if the solution is always the same within the last disc_consecutive_forbid
 """
 function check_solution_history(m::Optimizer, ind::Int)
     Alp.get_option(m, :disc_consecutive_forbid) == 0 && return false
-    (m.logs[:n_iter] < Alp.get_option(m, :disc_consecutive_forbid)) && return false
+    ((m.logs[:n_iter]-1) < Alp.get_option(m, :disc_consecutive_forbid)) && return false
 
     sol_val = m.bound_sol_history[mod(
-        m.logs[:n_iter] - 1,
+        m.logs[:n_iter] - 2,
         Alp.get_option(m, :disc_consecutive_forbid),
     )+1][ind]
     for i in 1:(Alp.get_option(m, :disc_consecutive_forbid)-1)
         search_pos =
-            mod(m.logs[:n_iter] - 1 - i, Alp.get_option(m, :disc_consecutive_forbid)) + 1
+            mod(m.logs[:n_iter] - 2 - i, Alp.get_option(m, :disc_consecutive_forbid)) + 1
         !isapprox(
             sol_val,
             m.bound_sol_history[search_pos][ind];